+++ Extract/Katakoi.h
@@ -21,6 +21,6 @@ private:
 	void GetBit(const u8*& src, u32& flags);
 	bool DecompImage(u8* dst, size_t dst_size, const u8* src, size_t src_size);
 
-	bool Compose(u8* dst, size_t dst_size, const u8* src, size_t src_size, long dst_width, long src_width, u16 bpp);
-	bool DecodeCompose(CArcFile* archive, const u8* diff, size_t diff_size, long diff_width, long diff_height, u16 diff_bpp);
+	bool Compose(u8* dst, size_t dst_size, const u8* src, size_t src_size, s32 dst_width, s32 src_width, u16 bpp);
+	bool DecodeCompose(CArcFile* archive, const u8* diff, size_t diff_size, s32 diff_width, s32 diff_height, u16 diff_bpp);
 };
